バグ修正２

+++ Assets/Scripts/TimingManager.cs
@@ -116,7 +116,7 @@ public class TimingManager : MonoBehaviour
         else if (spawn_enemy.Count > spawn_num)
         {
             float perfect_time = spawn_time[spawn_num];
-            if (Math.Abs(perfect_time - keyinput_time) <= judgment_width * (108 / database.BPM))
+            if (Math.Abs(perfect_time - keyinput_time) <= judgment_width * (108f / (float)database.BPM))
             {
                 if (KeyID == spawn_type[spawn_num] || (KeyID == 4 && spawn_type[spawn_num] == 2))
                 {
@@ -129,7 +129,7 @@ public class TimingManager : MonoBehaviour
                 spawn_num++;
             }
 
-            else if (Math.Abs(perfect_time - keyinput_time) <= 2f * judgment_width * (108 / database.BPM))
+            else if (Math.Abs(perfect_time - keyinput_time) <= 2f * judgment_width * (108f / (float)database.BPM))
             {
                 spawn_enemy[spawn_num].miss();
                 spawn_num++;
